+++ src/modules/DepositionPointCharge/DepositionPointChargeModule.cpp
@@ -134,9 +134,9 @@ void DepositionPointChargeModule::run(Event* event) {
                    ROOT::Math::XYZVector(
                        model->getPixelSize().x() / 2.0, model->getPixelSize().y() / 2.0, model->getSensorSize().z() / 2.0);
         LOG(DEBUG) << "Reference: " << ref;
-        position = ROOT::Math::XYZPoint(voxel_.x() * ((event->number - 1) % root_),
-                                        voxel_.y() * (((event->number - 1) / root_) % root_),
-                                        voxel_.z() * (((event->number - 1) / root_ / root_) % root_)) +
+        position = ROOT::Math::XYZPoint(voxel_.x() * static_cast<double>((event->number - 1) % root_),
+                                        voxel_.y() * static_cast<double>(((event->number - 1) / root_) % root_),
+                                        voxel_.z() * static_cast<double>(((event->number - 1) / root_ / root_) % root_)) +
                    ref;
     } else {
         // Calculate random offset from configured position
